+++ src/quacc/settings.py
@@ -375,6 +375,9 @@ class QuaccSettings(BaseSettings):
         ],
         description="Handlers for Custodian",
     )
+    VASP_CUSTODIAN_VASP_ERROR_EXCLUDE: list[str] = Field(
+        [], description="VASP errors to exclude from Custodian's VaspErrorHandler"
+    )
     VASP_CUSTODIAN_VALIDATORS: list[str] = Field(
         ["VasprunXMLValidator", "VaspFilesValidator"],
         description="Validators for Custodian",