+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project generally adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.7]
+
+### Added
+
+- MD recipes for MLIPs
+
+## [1.5.6]
+
+### Added
+
+- Added feature to ignore specific VASP custodian handlers via the quacc settings
+
+## [1.5.5]
+
+### Added
+
+- `omol_static_job` recipe
+
+### Fixed
+
+- Fixed bug where elastic workflows would always relax the cell
+
 ## [1.5.4]
 
 ### Added